requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
praw>=7.6.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
schedule>=1.1.0
pyyaml>=6.0
tenacity>=8.0.0
google-api-python-client>=2.0.0
youtube-transcript-api>=0.6.0
